requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
absl-py=0.13.0=pyhd8ed1ab_0
astroid=2.4.2=pypi_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
blessings=1.7=pypi_0
bzip2=1.0.8=h7b6447c_0
c-ares=1.17.1=h7f98852_1
ca-certificates=2021.5.30=ha878542_0
cairo=1.16.0=hf32fb01_1
certifi=2021.5.30=py36h5fab9bb_0
cffi=1.14.5=py36h261ae71_0
cuda100=1.0=0
cuda80=1.0=h205658b_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py_2
cython=0.29.24=py36h295c915_0
dataclasses=0.8=pyh787bdff_0
dbus=1.13.6=he372182_0
easydict=1.9=py_0
expat=2.2.10=he6710b0_2
faiss=1.6.3=py36hb679d2d_1_cuda
faiss-cpu=1.6.3=py36h6bb024c_0
faiss-gpu=1.6.3=py36h1a5d453_0
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.1=hba837de_1004
freeglut=3.0.0=hf484d3e_5
freetype=2.10.4=h5ab3b9f_0
future=0.18.2=pypi_0
glib=2.67.4=h36276a3_1
gpustat=0.6.0=pypi_0
graphite2=1.3.14=h23475e2_0
grpcio=1.33.2=py36he0f7d3b_2
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=h28cd5cc_2
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=hf484d3e_1000
importlib-metadata=4.6.2=py36h5fab9bb_0
inplace-abn=1.1.1.dev3+ga8a3f82=pypi_0
intel-openmp=2020.2=254
isort=5.5.4=pypi_0
jasper=2.0.14=h07fcdf6_1
joblib=0.17.0=pypi_0
jpeg=9b=habf39ab_1
kiwisolver=1.3.1=py36h51d7077_0
lazy-object-proxy=1.4.3=pypi_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libblas=3.8.0=16_mkl
libedit=3.1.20191231=h14c3975_1
libfaiss=1.6.3=he61ee18_1_cuda
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h24d8f2e_14
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1
libgomp=9.3.0=h24d8f2e_14
liblapack=3.8.0=16_mkl
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=2.32.1=h7f98852_1000
libvpx=1.7.0=h439df22_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lz4-c=1.9.2=heb0550a_3
markdown=3.3.4=pyhd8ed1ab_0
matplotlib=3.3.4=py36h5fab9bb_0
matplotlib-base=3.3.4=py36h62a2d02_0
mccabe=0.6.1=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.2.0=py36h23d657b_0
mkl_random=1.1.1=py36h0573a6f_0
ml-metrics=0.1.4=pypi_0
mllogger=0.7=pypi_0
ncurses=6.2=he6710b0_1
ninja=1.10.2=py36hff7bd54_0
numpy=1.19.1=py36hbc911f0_0
numpy-base=1.19.1=py36hfa32c7d_0
nvidia-ml-py3=7.352.0=pypi_0
olefile=0.46=py36_0
opencv=3.4.2=py36h6fd60c2_1
openssl=1.1.1k=h7f98852_0
pandas=1.1.5=pypi_0
pcre=8.44=he1b5a44_0
pillow=8.0.0=py36h9a89aac_0
pip=20.2.3=py36_0
pixman=0.40.0=h7b6447c_0
prettytable=2.1.0=pyhd8ed1ab_0
protobuf=3.17.3=pypi_0
psutil=5.7.2=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
py-opencv=3.4.2=py36hb342d67_1
pycocotools=2.0.2=py36h4d9540e_2
pycparser=2.20=py_2
pylint=2.6.0=pypi_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.9.2=py36hcca6a23_4
python=3.6.10=h7579374_2
python-dateutil=2.8.1=py_0
python-lmdb=1.2.1=py36h2531618_1
python_abi=3.6=1_cp36m
pytorch=1.6.0=py3.6_cuda10.2.89_cudnn7.6.5_0
pytz=2021.1=pypi_0
qt=5.9.7=h5867ecd_1
randaugment=1.0.2=pypi_0
readline=8.0=h7b6447c_0
scikit-learn=0.23.2=pypi_0
scipy=1.5.2=pypi_0
setuptools=50.3.0=py36hb0f4dca_1
sip=4.19.8=py36hf484d3e_0
six=1.15.0=py_0
sklearn=0.0=pypi_0
sqlite=3.33.0=h62c20be_0
tensorboard=1.15.0=py36_0
tensorboardx=2.2=pyhd8ed1ab_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=hbc83047_0
toml=0.10.1=pypi_0
torchvision=0.2.2=py_3
tornado=6.1=py36h8f6f2f9_1
tqdm=4.50.0=pypi_0
typed-ast=1.4.1=pypi_0
typing_extensions=3.10.0.0=pyha770c72_0
wcwidth=0.2.5=pyh9f0ad1d_2
werkzeug=2.0.1=pyhd8ed1ab_0
wheel=0.35.1=py_0
wrapt=1.12.1=pypi_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.5=h7b6447c_0
zipp=3.5.0=pyhd8ed1ab_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.4=h0b5b093_3
